created the module market (This commit doesn't run!)
